Stop checking consistency of babble cache
There's probably some kind of bug here, but it's realistically not going
to get fixed, and missing a line or two from the babble cache isn't
exactly the end of the world.

e.g.
02:56:32 msbobBot | Error in channel #msbobbot-control -- the bot --
Exception in workers.py on line 166: Last row in babble cache (2445683) does not match last row in log (2445686).